+++ router/router.go
@@ -8,8 +8,8 @@ import (
 func SetRouters(m *macaron.Macaron) {
 
 	m.Get("/", handler.IndexHandler)
-	m.Get("/ws", handler.ServeWs)
-	m.Get("/ws_getbuildlog", handler.WSGetLog)
-	m.Post("/show", handler.GetLog)
-	m.Post("/buildreq", handler.SendBuildReq)
+	//m.Get("/ws", handler.WSServer)
+	m.Get("/wsbuildlog", handler.WSbuildLog)
+	m.Post("/httpbuildlog", handler.HTTPBuildLog)
+	m.Post("/httpbuild", handler.HTTPBuild)
 }
